refactor: move recent algorithms into explore dropdown
