+++ src/components/Navbar.jsx
@@ -1,4 +1,4 @@
-import { useState } from 'react'
+import { useState, useEffect } from 'react'
 import { Link, useLocation } from 'react-router-dom'
 import {
   SignedIn,
@@ -6,61 +6,58 @@ import {
   SignInButton,
   UserButton,
 } from '@clerk/clerk-react'
-// 1. Import motion and AnimatePresence
 import { motion, AnimatePresence } from 'framer-motion'
+
 import githubIcon from '../assets/github-mark-white.svg'
 import logo from '../assets/logo2.png'
 import SearchBar from './SearchBar'
 
-// 2. Define the bounce transition
 const bounceTransition = {
   type: 'spring',
   stiffness: 260,
   damping: 15,
 }
 
-// 3. Define variants for the icon lines
 const topVariants = {
   closed: { rotate: 0, y: 0 },
-  open: { rotate: 45, y: 6 }, // Move down 6px
+  open: { rotate: 45, y: 6 },
 }
+
 const middleVariants = {
   closed: { opacity: 1 },
   open: { opacity: 0 },
 }
+
 const bottomVariants = {
   closed: { rotate: 0, y: 0 },
-  open: { rotate: -45, y: -6 }, // Move up 6px
+  open: { rotate: -45, y: -6 },
 }
 
-// 4. Define variants for the mobile menu panel
 const menuVariants = {
   closed: {
     opacity: 0,
-    y: -10, // Start 10px up
+    y: -10,
     transition: {
       duration: 0.2,
     },
   },
   open: {
     opacity: 1,
-    y: 0, // Animate to 0
+    y: 0,
     transition: {
       type: 'spring',
       stiffness: 260,
       damping: 20,
-      staggerChildren: 0.05, // Animate links one by one
+      staggerChildren: 0.05,
     },
   },
 }
 
-// Variant for the links inside the mobile menu
 const menuItemVariants = {
   closed: { opacity: 0, y: -10 },
   open: { opacity: 1, y: 0 },
 }
 
-// Helper component for the icon lines
 const Line = ({ variants }) => (
   <motion.div
     className="h-0.5 w-5 bg-slate-300"
@@ -71,8 +68,18 @@ const Line = ({ variants }) => (
 
 export const Navbar = () => {
   const [open, setOpen] = useState(false)
-  // Derive active state from current URL instead of local state
+
   const { pathname } = useLocation()
+
+  const [history, setHistory] = useState(() => {
+    try {
+      const saved = localStorage.getItem('algo-history')
+      return saved ? JSON.parse(saved) : []
+    } catch (error) {
+      console.error('Failed to parse algo-history:', error)
+      return []
+    }
+  })
 
   const algorithmLinks = [
     { name: 'Search', href: '/search' },
@@ -83,6 +90,22 @@ export const Navbar = () => {
     { name: "Kadane's Algorithm", href: '/kadane' },
     { name: "Moore's Voting Algorithm", href: '/moore-voting' },
   ]
+
+  useEffect(() => {
+    const current = algorithmLinks.find((link) => link.href === pathname)?.name
+
+    if (current) {
+      setHistory((prev) => {
+        const updated = [current, ...prev.filter((item) => item !== current)]
+
+        const trimmed = updated.slice(0, 5)
+
+        localStorage.setItem('algo-history', JSON.stringify(trimmed))
+
+        return trimmed
+      })
+    }
+  }, [pathname])
 
   return (
     <header className="sticky top-0 z-50 w-full border-b border-white/10 bg-slate-950/50 backdrop-blur supports-[backdrop-filter]:bg-slate-950/40 rounded-xl shadow-2xl">
@@ -95,6 +118,7 @@ export const Navbar = () => {
             <div className="w-10 h-10 m-auto rounded flex items-center justify-center mr-3 transition-transform group-hover:scale-110">
               <img src={logo} alt="AlgoScope Logo" className="w-8 h-8" />
             </div>
+
             <span className="mt-1 text-2xl text-white font-bold tracking-tighter">
               AlgoScope
             </span>
@@ -112,7 +136,7 @@ export const Navbar = () => {
                   Explore
                 </button>
 
-                <div className="absolute left-0 top-12 py-2 invisible opacity-0 translate-y-2 min-w-[220px] rounded-xl border border-white/10 bg-slate-900/95 p-2 shadow-2xl backdrop-blur-xl transition-all duration-200 group-hover:visible group-hover:opacity-100 group-hover:translate-y-0 z-50">
+                <div className="absolute left-0 top-12 py-2 invisible opacity-0 translate-y-2 min-w-[240px] rounded-xl border border-white/10 bg-slate-900/95 p-2 shadow-2xl backdrop-blur-xl transition-all duration-200 group-hover:visible group-hover:opacity-100 group-hover:translate-y-0 z-50">
                   {algorithmLinks.map((link) => (
                     <Link
                       key={link.name}
@@ -126,9 +150,38 @@ export const Navbar = () => {
                       {link.name}
                     </Link>
                   ))}
+
+                  <div className="my-2 border-t border-white/10" />
+
+                  <p className="px-4 py-2 text-xs font-semibold uppercase tracking-widest text-slate-500">
+                    Recent
+                  </p>
+
+                  {history.length === 0 ? (
+                    <p className="px-4 py-2 text-sm text-slate-500">
+                      No recent algorithms
+                    </p>
+                  ) : (
+                    history.map((item) => {
+                      const matched = algorithmLinks.find(
+                        (link) => link.name === item
+                      )
+
+                      return (
+                        <Link
+                          key={item}
+                          to={matched?.href || '/'}
+                          className="block rounded-lg px-4 py-2 text-sm text-slate-300 hover:bg-white/5 hover:text-white transition-all"
+                        >
+                          {item}
+                        </Link>
+                      )
+                    })
+                  )}
                 </div>
               </li>
             </ul>
+
             <Link
               to="https://github.com/algoscope-hq/AlgoScope"
               className="inline-flex items-center rounded-xl bg-white px-5 py-2 text-sm font-bold text-black shadow-lg hover:bg-slate-200 transition-all duration-200 active:scale-95"
@@ -138,6 +191,7 @@ export const Navbar = () => {
                 alt="Github Repository Link"
                 className="w-7 h-5 pr-2 invert"
               />
+
               <span>Github</span>
             </Link>
 
@@ -146,10 +200,12 @@ export const Navbar = () => {
                 <SignInButton mode="modal">
                   <button className="relative group overflow-hidden rounded-xl bg-slate-900 px-6 py-2 text-sm font-bold text-white transition-all duration-300 active:scale-95">
                     <span className="relative z-10">Sign In</span>
+
                     <div className="absolute inset-0 bg-gradient-to-r from-indigo-600/20 to-purple-600/20 opacity-0 group-hover:opacity-100 transition-opacity duration-300" />
                   </button>
                 </SignInButton>
               </SignedOut>
+
               <SignedIn>
                 <UserButton
                   appearance={{
@@ -163,7 +219,7 @@ export const Navbar = () => {
             </div>
           </div>
 
-          {/* 5. Apply the animation to the button */}
+          {/* Mobile Menu Button */}
           <div className="flex items-center gap-4 md:hidden">
             <SignedIn>
               <UserButton
@@ -174,16 +230,15 @@ export const Navbar = () => {
                 }}
               />
             </SignedIn>
+
             <motion.button
               type="button"
               aria-label="Toggle menu"
               aria-expanded={open}
               onClick={() => setOpen((o) => !o)}
-              // Animate between 'open' and 'closed' states
               animate={open ? 'open' : 'closed'}
               className="inline-flex flex-col items-center justify-center gap-1 rounded-lg p-2 text-slate-300 hover:bg-white/10 focus:outline-none focus:ring-2 focus:ring-white/20 transition-colors"
             >
-              {/* 6. Remove old SVGs and add animated lines */}
               <Line variants={topVariants} />
               <Line variants={middleVariants} />
               <Line variants={bottomVariants} />
@@ -192,27 +247,24 @@ export const Navbar = () => {
         </div>
       </div>
 
-      {/* 7. Animate the mobile menu dropdown */}
+      {/* Mobile Menu */}
       <AnimatePresence>
         {open && (
           <motion.div
             key="mobile-menu"
-            // Remove Tailwind's show/hide, let framer-motion handle it
             className="md:hidden border-t border-white/5 bg-slate-950/90 backdrop-blur-xl shadow-2xl rounded-b-2xl overflow-hidden"
-            // Apply variants
             variants={menuVariants}
             initial="closed"
             animate="open"
             exit="closed"
           >
             <div className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8 py-6">
-              {/* Mobile Search */}
               <div className="mb-6 lg:hidden">
                 <SearchBar />
               </div>
+
               <ul className="space-y-2">
                 {algorithmLinks.map((link) => (
-                  // Animate each link
                   <motion.li key={link.name} variants={menuItemVariants}>
                     <Link
                       to={link.href}
@@ -237,10 +289,12 @@ export const Navbar = () => {
                   <SignInButton mode="modal">
                     <button className="w-full relative group overflow-hidden rounded-xl bg-slate-900 border border-white/10 px-4 py-3 text-base font-bold text-white transition-all duration-300 active:scale-[0.98]">
                       <span className="relative z-10">Sign In</span>
+
                       <div className="absolute inset-0 bg-gradient-to-r from-indigo-600/10 to-purple-600/10" />
                     </button>
                   </SignInButton>
                 </SignedOut>
+
                 <Link
                   to="https://github.com/algoscope-hq/AlgoScope"
                   onClick={() => setOpen(false)}
